+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Changelog
   simulated in V-REP.
 - Added setting object orientation via interface to a generic scene object
   simulated in V-REP.
+- Added retrieving limits of object bounding box via interface to a generic
+  scene object simulated in V-REP.
 - Added retrieving limits of model bounding box via interface to a generic
   model simulated in V-REP.
 - Added copying and pasting object via interface to a generic scene object